retrieving the userID for the event creation

+++ models/event.go
@@ -9,12 +9,12 @@ import (
 // Event represents an event with ID, name, description, location, date/time, and user ID fields.
 // The struct fields are validated as required.
 type Event struct {
-	ID          int64     
+	ID          int64
 	Name        string    `binding:"required"`
 	Description string    `binding:"required"`
 	Location    string    `binding:"required"`
 	DateTime    time.Time `binding:"required"`
-	UserID      int       `binding:"required"`
+	UserID      int64     `binding:"required"`
 }
 
 // Save inserts the Event into the database.
